requirements.txt
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
Pillow==8.2.0
Werkzeug==2.0.1
